Lightstep/GlobalTracer compat
